python/requirements.txt
pynacl==1.5.0
canonicaljson==2.0.0